// frontend/src/app/core/models/filial.model.ts
export interface Filial {
    id?: string;
    razaoSocial: string;
    nomeFantasia?: string;
    cnpj: string;
    inscricaoEstadual?: string;
    inscricaoMunicipal?: string;
    crt?: string;
    tenantIdentifier?: string;
    cep?: string;
    logradouro?: string;
    numero?: string;
    complemento?: string;
    bairro?: string;
    cidade?: string;
    estado?: string;
    ibge?: string;
    isMatriz?: boolean;
    ativo?: boolean;
}